Small change for master+slaves update
As an example, I'm using a remote setup. 1 master and 2 machine slaves with 3 nodes each. In my slaves machines the user for the nodes is "elrond" and the $CUSTOMHOME is "/home/elrond". But in my master machine it isn't the case, I'm using other user with other home. I think that if you use in this case "relative" aproach, like create a brother folder it will work always.